Restructure of function/order and readme change
Ambiguous function names such as "addTab" has now been clarified, using the word "monitor" to indicate whether a tab is being monitored or added initially as a potential tab.

genTabList no longer creates the document list, it now only requests and passes the list to renderTabList to mirror the functionality of monitoring tabs.